MVC for lists and bookmarks

+++ db/seeds.rb
@@ -7,7 +7,11 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-Movie.destroy_all
+if Rails.env.develoment?
+  Movie.destroy_all
+  List.destroy_all
+end
+
 Movie.create(title: 'Wonder Woman 1984',
              overview: 'Wonder Woman comes into conflict with the Soviet Union during the Cold War in the 1980s', poster_url: 'https://image.tmdb.org/t/p/original/8UlWHLMpgZm9bx6QYh0NFoq67TZ.jpg', rating: 6.9)
 Movie.create(title: 'The Shawshank Redemption',
@@ -17,7 +21,7 @@ Movie.create(title: 'Titanic',
 Movie.create(title: "Ocean's Eight",
              overview: 'Debbie Ocean, a criminal mastermind, gathers a crew of female thieves to pull off the heist of the century.', poster_url: 'https://image.tmdb.org/t/p/original/MvYpKlpFukTivnlBhizGbkAe3v.jpg', rating: 7.0)
 
-List.destroy_all
+
 List.create(name: 'horror',
             bg_url: 'https://images.unsplash.com/photo-1601513445506-2ab0d4fb4229?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=668&q=80')
 List.create(name: 'Sci-Fi',
